Move the Playwright console error detection to fixtures. Start writing some docs about all this container/decorator stuff.
